initialize with bones of inverse gain
This bindgen approach is not working
The wasm loading has to be lazy/async,
and there's no way to fetch from the audio thread.
This needs to be packaged absolutely raw.
This version seems to have dependencies from
the wasm file to the js wrapper somehow
and it fails to initialize